It's been a pain but I was able to finally get data from the mysql server, sent to my front end vue components. Now I just need to figure out how to have updates and things like that in the database based on the actions of the user. This has been a good learning experience.
I am going to have to figure out how to get all my scripts to run on the backend server, as having it in the front end is a massive security risk in my opinion. I may never push these files to a website, however, taking precautions now can't be a bad idea right. Also it will teach me how to even have backend scripts be run.

I will need to start working on the collection system as well so that when a user opens a pack it will be displayed in their account. Just a little bit more satisfaction on every pull.

Currently, working on a way to turn this into more than just a tradiing card game. I'm, thinking of turning this into a turn based, spell type of game where the packs can be used to pull gear for your character and your minions. This is a long way out of the picture considering what I'm struggling with right now, but It's always good to dream big.